feat: modularity and use of controller
