+++ Dockerfile
@@ -6,8 +6,8 @@ COPY package*.json .
 
 RUN npm install
 
-COPY . ./
+COPY ./src ./
 
 EXPOSE 8080
 
-CMD [ "npm","start" ]
+CMD [ "npm","run","start" ]
