+++ src/userCases/getUsers/getUsers.js
@@ -1,0 +1,12 @@
+const User = require('../../models/user')
+
+async function getUser() {
+    try {
+        const users = await User.find()
+        return(users)
+    } catch (error) {
+        return(error)
+    }
+}
+
+module.exports = getUser
